+++ firmware/prototyper/src/driver/clint/sifive.rs
@@ -177,7 +177,7 @@ impl SiFiveIpi {
 
 impl IpiBackend for SiFiveIpi {
     #[inline(always)]
-    fn send_ipi(&mut self, req: IpiRequest) -> Result<(), IpiError> {
+    fn send_ipi(&self, req: IpiRequest) -> Result<(), IpiError> {
         for hart_id in req.harts() {
             self.write(IpiRegister::Msip, hart_id, IpiState::Pending)?;
         }
@@ -185,7 +185,7 @@ impl IpiBackend for SiFiveIpi {
     }
 
     #[inline(always)]
-    fn clear_ipi(&mut self, hart_id: usize) -> Result<(), IpiError> {
+    fn clear_ipi(&self, hart_id: usize) -> Result<(), IpiError> {
         self.write(IpiRegister::Msip, hart_id, IpiState::Clear)
     }
 }